Outsourced filtering profile for public fields
- Moved profile utils.server from /settings folder to /$username folder
- Moved deriveMode function inside utils.server
- Created filterProfileByMode(profile, mode) inside utils.server
- Created transformEventData() inside utils.server to make code inside filterProfileByMode() clearer
