+++ app/routes/profile/$username/settings/delete.tsx
@@ -10,7 +10,7 @@ import {
   getOrganisationsOnProfileByUserId,
   getProfileByUserId,
 } from "~/profile.server";
-import { handleAuthorization } from "./utils.server";
+import { handleAuthorization } from "../utils.server";
 
 const schema = z.object({
   id: z.string().uuid(),
